requirements.txt
asgiref==3.9.2
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
PyMySQL==1.1.2
python-decouple==3.8
sqlparse==0.5.3
